backend/requirements.txt
# WashU Navigator Backend
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.28.1

# PDF processing
pymupdf==1.24.0
pdfplumber==0.11.0

# Vectorless RAG ΓÇö E10's PageIndex fork
google-genai==1.7.0
google-cloud-aiplatform==1.60.0
PyYAML==6.0.1

# Data / storage
sqlalchemy==2.0.30
aiosqlite==0.20.0

# Utils
python-multipart==0.0.9
